% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airlines.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline names}
\format{
Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation.}
\item{name}{Full name.}
}
}
\usage{
airlines
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  airlines
}
\keyword{datasets}
